+++ client/web/src/site-admin/SiteAdminAllUsersPage/UserManagement/components/ActivityChart.tsx
